requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
